Change input management to the main
