--- SturdyMachine/Assets/Resources/Configuration/OffenseBlockingConfig.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 47d40567136b2284a8f5c8847b6af5f1, type: 3}
-  m_Name: OffenseBlockingConfig
-  m_EditorClassIdentifier: 
-  _offenseBlocking:
-  - {fileID: 11400000, guid: f1f4199aea9fb624d95b14511d9ff73e, type: 2}
