--- SturdyMachine/Assets/Resources/Configuration/OffenseBlockingConfig.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c0807c0d0b20e7f438e3ead6e8819e39
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
