+++ SturdyMachine/Assets/Scripts/Manager/Main.cs
@@ -1,4 +1,6 @@
 ﻿using UnityEngine;
+using UnityEngine.InputSystem;
+using UnityEngine.InputSystem.Interactions;
 
 #if UNITY_EDITOR
 using UnityEditor;
@@ -7,43 +9,300 @@ using UnityEditor;
 using ICustomEditor.Class;
 using Feature.Manager;
 
+using Humanoid.Bot.Sturdy;
+
 namespace GameplayFeature.Manager
 {
     [RequireComponent(typeof(FeatureManager))]
     public class Main : UnityICustomEditor
     {
+        //Sturdy
         [SerializeField]
-        Transform _sturdyMachine;
+        SturdyMachineControls _sturdyMachineControl;
+
+        [SerializeField]
+        SturdyBot _sturdyBot;
+
+        OffenseDirection _currentOffenseDirection;
+        OffenseType _currentOffenseType;
+
+        bool _isStanceActivated;
 
         [SerializeField]
         FeatureManager _featureManager;
 
-        static Main _main;
-
-        public static Main GetInstance => _main;
         public FeatureManager GetFeatureManager { get { return _featureManager; } }
-        public Transform GetSturdyMachine => _sturdyMachine;
 
         public override void Awake()
         {
-            _main = this;
+            //Sturdy
+            _sturdyMachineControl = new SturdyMachineControls();
+
+            OffenseStanceSetup();
+
+            _sturdyBot.Awake();
 
             _featureManager.Awake();
         }
 
         public override void Start() 
         {
+            _currentOffenseDirection = OffenseDirection.STANCE;
+            _currentOffenseType = OffenseType.DEFAULT;
+
             _featureManager.Start();
         }
 
         void Update()
         {
-            _featureManager.CustomUpdate(_sturdyMachine.position);
+            SturdySetup();
+
+            _featureManager.CustomUpdate(_sturdyBot.transform.position);
         }
 
         void LateUpdate()
         {
+            _sturdyBot.LateUpdateRemote(_currentOffenseDirection);
+
             _featureManager.LateUpdate();
+        }
+
+        void OnEnable()
+        {
+            _sturdyMachineControl.Deflection.Enable();
+            _sturdyMachineControl.Sweep.Enable();
+            _sturdyMachineControl.Strikes.Enable();
+            _sturdyMachineControl.Heavy.Enable();
+            _sturdyMachineControl.DeathBlow.Enable();
+        }
+
+        void OnDisable()
+        {
+            _sturdyMachineControl.Deflection.Disable();
+            _sturdyMachineControl.Sweep.Disable();
+            _sturdyMachineControl.Strikes.Disable();
+            _sturdyMachineControl.Heavy.Disable();
+            _sturdyMachineControl.DeathBlow.Disable();
+        }
+
+        void OffenseStanceSetup()
+        {
+            #region Strikes
+
+            _sturdyMachineControl.Strikes.Strikes.performed += context =>
+            {
+                if (context.interaction is HoldInteraction)
+                {
+                    if (!_isStanceActivated)
+                        _isStanceActivated = true;
+                }
+                else if (_isStanceActivated)
+                    _isStanceActivated = false;
+
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+
+                    if (_currentOffenseType != OffenseType.STRIKE)
+                        _currentOffenseType = OffenseType.STRIKE;
+                }
+            };
+
+            _sturdyMachineControl.Strikes.Strikes.canceled += context =>
+            {
+                if (_isStanceActivated)
+                    _isStanceActivated = false;
+            };
+
+            #endregion
+
+            #region Heavy
+
+            _sturdyMachineControl.Heavy.Heavy.performed += context =>
+            {
+                if (context.interaction is HoldInteraction)
+                {
+                    if (!_isStanceActivated)
+                        _isStanceActivated = true;
+                }
+                else if (_isStanceActivated)
+                    _isStanceActivated = false;
+
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+
+                    if (_currentOffenseType != OffenseType.HEAVY)
+                        _currentOffenseType = OffenseType.HEAVY;
+                }
+            };
+
+            _sturdyMachineControl.Heavy.Heavy.canceled += context =>
+            {
+                if (_isStanceActivated)
+                    _isStanceActivated = false;
+            };
+
+            #endregion
+
+            #region DeathBlow
+
+            _sturdyMachineControl.DeathBlow.DeathBlow.performed += context =>
+            {
+                if (context.interaction is HoldInteraction)
+                {
+                    if (!_isStanceActivated)
+                        _isStanceActivated = true;
+                }
+                else if (_isStanceActivated)
+                    _isStanceActivated = false;
+
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+
+                    if (_currentOffenseType != OffenseType.DEATHBLOW)
+                        _currentOffenseType = OffenseType.DEATHBLOW;
+                }
+            };
+
+            _sturdyMachineControl.DeathBlow.DeathBlow.canceled += context =>
+            {
+                if (_isStanceActivated)
+                    _isStanceActivated = false;
+
+            };
+
+            #endregion
+
+            #region Deflection
+
+            //Neutral
+            _sturdyMachineControl.Deflection.Neutral.performed += context =>
+            {
+                if (context.interaction is PressInteraction)
+                {
+                    _currentOffenseDirection = OffenseDirection.NEUTRAL;
+
+                    if (!_isStanceActivated)
+                    {
+                        if (_currentOffenseType != OffenseType.DEATHBLOW)
+                            _currentOffenseType = OffenseType.DEFLECTION;
+                    }
+                }
+            };
+
+            _sturdyMachineControl.Deflection.Neutral.canceled += context =>
+            {
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+                }
+            };
+
+            //Right
+            _sturdyMachineControl.Deflection.Right.performed += context =>
+            {
+                if (context.interaction is PressInteraction)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.RIGHT)
+                        _currentOffenseDirection = OffenseDirection.RIGHT;
+
+                    if (!_isStanceActivated)
+                    {
+                        if (_currentOffenseType != OffenseType.DEFLECTION)
+                            _currentOffenseType = OffenseType.DEFLECTION;
+                    }
+                }
+            };
+
+            _sturdyMachineControl.Deflection.Right.canceled += context =>
+            {
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+                }
+            };
+
+            //Left
+            _sturdyMachineControl.Deflection.Left.performed += context =>
+            {
+                if (context.interaction is PressInteraction)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.LEFT)
+                        _currentOffenseDirection = OffenseDirection.LEFT;
+
+                    if (!_isStanceActivated)
+                    {
+                        if (_currentOffenseType != OffenseType.DEFLECTION)
+                            _currentOffenseType = OffenseType.DEFLECTION;
+                    }
+                };
+            };
+
+            _sturdyMachineControl.Deflection.Left.canceled += context =>
+            {
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+                }
+            };
+
+            //Evasion
+            _sturdyMachineControl.Deflection.Evasion.performed += context =>
+            {
+                if (context.interaction is PressInteraction)
+                {
+                    if (_isStanceActivated)
+                    {
+                        if (_currentOffenseDirection != OffenseDirection.NEUTRAL)
+                            _currentOffenseDirection = OffenseDirection.NEUTRAL;
+
+                        if (_currentOffenseType != OffenseType.SWEEP)
+                            _currentOffenseType = OffenseType.SWEEP;
+
+                    }
+                    else
+                    {
+                        if (_currentOffenseDirection != OffenseDirection.NEUTRAL)
+                            _currentOffenseDirection = OffenseDirection.NEUTRAL;
+
+                        if (_currentOffenseType != OffenseType.EVASION)
+                            _currentOffenseType = OffenseType.EVASION;
+                    }
+                }
+            };
+
+            _sturdyMachineControl.Deflection.Evasion.canceled += context =>
+            {
+                if (_isStanceActivated)
+                {
+                    if (_currentOffenseDirection != OffenseDirection.STANCE)
+                        _currentOffenseDirection = OffenseDirection.STANCE;
+                }
+            };
+
+            #endregion
+        }
+
+        void SturdySetup() 
+        {
+            _sturdyBot.UpdateRemote(_currentOffenseDirection, _currentOffenseType, _isStanceActivated);
+
+            if (!_isStanceActivated)
+            {
+                if (_currentOffenseType != OffenseType.DEFAULT)
+                    _currentOffenseType = OffenseType.DEFAULT;
+            }
+
+            if (_currentOffenseDirection != OffenseDirection.STANCE)
+                _currentOffenseDirection = OffenseDirection.STANCE;
         }
 
 #if UNITY_EDITOR
@@ -63,7 +322,7 @@ namespace GameplayFeature.Manager
 
             EditorGUILayout.BeginVertical(GUI.skin.box);
 
-            _sturdyMachine = (Transform)EditorGUILayout.ObjectField("SturdyMachine:", _sturdyMachine, typeof(Transform), true);
+            _sturdyBot = EditorGUILayout.ObjectField("SturdyBot:", _sturdyBot, typeof(SturdyBot), true) as SturdyBot;
 
             EditorGUILayout.EndVertical();
