+++ SturdyMachine/Assets/Scripts/Player/SturdyBot.cs
@@ -14,274 +14,28 @@ namespace Humanoid.Bot.Sturdy
     [RequireComponent(typeof(PlayerInput))]
     public class SturdyBot : Bot
     {
-        SturdyMachineControls _sturdyMachineControl;
-
-        bool _isStanceActivated;
-
-        OffenseDirection _currentOffenseDirection;
-        OffenseType _currentOffenseType;
-
         public override void Awake()
         {
             base.Awake();
-
-            _sturdyMachineControl = new SturdyMachineControls();
-
-            OffenseStanceSetup();
         }
 
         public override void Start()
         {
             base.Start();
-
-            _currentOffenseDirection = OffenseDirection.STANCE;
-            _currentOffenseType = OffenseType.DEFAULT;
-
         }
 
-        void Update() 
+        public virtual void UpdateRemote(OffenseDirection pOffenseDirection, OffenseType pOffenseType, bool pIsStanceActivated) 
         {
-            base.CustomUpdate(_currentOffenseDirection, _currentOffenseType, _isStanceActivated);
-
-            if (!_isStanceActivated)
-            {
-                if (_currentOffenseType != OffenseType.DEFAULT)
-                    _currentOffenseType = OffenseType.DEFAULT;
-            }
-
-            if (_currentOffenseDirection != OffenseDirection.STANCE)
-                _currentOffenseDirection = OffenseDirection.STANCE;
+            base.CustomUpdate(pOffenseDirection, pOffenseType, pIsStanceActivated);
 
             ////Focus
             //if (Main.GetInstance.GetFeatureManager.GetFocusManager.GetCurrentFocus)
             //    transform.rotation = Quaternion.Slerp(transform.rotation, Quaternion.LookRotation(Main.GetInstance.GetFeatureManager.GetFocusManager.GetCurrentFocus.transform.position - transform.position), 0.07f);
         }
 
-        void LateUpdate() 
+        public virtual void LateUpdateRemote(OffenseDirection pOffenseDirection) 
         {
-            base.CustomLateUpdate(_currentOffenseDirection);
-        }
-
-        void OffenseStanceSetup() 
-        {
-            #region Strikes
-
-            _sturdyMachineControl.Strikes.Strikes.performed += context =>
-            {
-                if (context.interaction is HoldInteraction)
-                {
-                    if (!_isStanceActivated)
-                        _isStanceActivated = true;
-                }
-                else if (_isStanceActivated)
-                    _isStanceActivated = false;
-
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-
-                    if (_currentOffenseType != OffenseType.STRIKE)
-                        _currentOffenseType = OffenseType.STRIKE;
-                }
-            };
-
-            _sturdyMachineControl.Strikes.Strikes.canceled += context =>
-            {
-                if (_isStanceActivated)
-                    _isStanceActivated = false;
-            };
-
-            #endregion
-
-            #region Heavy
-
-            _sturdyMachineControl.Heavy.Heavy.performed += context =>
-            {
-                if (context.interaction is HoldInteraction)
-                {
-                    if (!_isStanceActivated)
-                        _isStanceActivated = true;
-                }
-                else if (_isStanceActivated)
-                    _isStanceActivated = false;
-
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-
-                    if (_currentOffenseType != OffenseType.HEAVY)
-                        _currentOffenseType = OffenseType.HEAVY;
-                }
-            };
-
-            _sturdyMachineControl.Heavy.Heavy.canceled += context =>
-            {
-                if (_isStanceActivated)
-                    _isStanceActivated = false;
-            };
-
-            #endregion
-
-            #region DeathBlow
-
-            _sturdyMachineControl.DeathBlow.DeathBlow.performed += context =>
-            {
-                if (context.interaction is HoldInteraction)
-                {
-                    if (!_isStanceActivated)
-                        _isStanceActivated = true;
-                }
-                else if (_isStanceActivated)
-                    _isStanceActivated = false;
-
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-
-                    if (_currentOffenseType != OffenseType.DEATHBLOW)
-                        _currentOffenseType = OffenseType.DEATHBLOW;
-                }
-            };
-
-            _sturdyMachineControl.DeathBlow.DeathBlow.canceled += context =>
-            {
-                if (_isStanceActivated)
-                    _isStanceActivated = false;
-
-            };
-
-            #endregion
-
-            /*---DEFLECTION---*/
-
-            //Neutral
-            _sturdyMachineControl.Deflection.Neutral.performed += context =>
-            {
-                if (context.interaction is PressInteraction)
-                {
-                    _currentOffenseDirection = OffenseDirection.NEUTRAL;
-
-                    if (!_isStanceActivated)
-                    {
-                        if (_currentOffenseType != OffenseType.DEATHBLOW)
-                            _currentOffenseType = OffenseType.DEFLECTION;
-                    }
-                }
-            };
-
-            _sturdyMachineControl.Deflection.Neutral.canceled += context =>
-            {
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-                }
-            };
-
-            //Right
-            _sturdyMachineControl.Deflection.Right.performed += context =>
-            {
-                if (context.interaction is PressInteraction)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.RIGHT)
-                        _currentOffenseDirection = OffenseDirection.RIGHT;
-
-                    if (!_isStanceActivated)
-                    {
-                        if (_currentOffenseType != OffenseType.DEFLECTION)
-                            _currentOffenseType = OffenseType.DEFLECTION;
-                    }
-                }
-            };
-
-            _sturdyMachineControl.Deflection.Right.canceled += context =>
-            {
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-                }
-            };
-
-            //Left
-            _sturdyMachineControl.Deflection.Left.performed += context =>
-            {
-                if (context.interaction is PressInteraction)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.LEFT)
-                        _currentOffenseDirection = OffenseDirection.LEFT;
-
-                    if (!_isStanceActivated)
-                    {
-                        if (_currentOffenseType != OffenseType.DEFLECTION)
-                            _currentOffenseType = OffenseType.DEFLECTION;
-                    }
-                };
-            };
-
-            _sturdyMachineControl.Deflection.Left.canceled += context =>
-            {
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-                }
-            };
-
-            //Evasion
-            _sturdyMachineControl.Deflection.Evasion.performed += context =>
-            {
-                if (context.interaction is PressInteraction)
-                {
-                    if (_isStanceActivated)
-                    {
-                        if (_currentOffenseDirection != OffenseDirection.NEUTRAL)
-                            _currentOffenseDirection = OffenseDirection.NEUTRAL;
-
-                        if (_currentOffenseType != OffenseType.SWEEP)
-                            _currentOffenseType = OffenseType.SWEEP;
-
-                    }
-                    else
-                    {
-                        if (_currentOffenseDirection != OffenseDirection.NEUTRAL)
-                            _currentOffenseDirection = OffenseDirection.NEUTRAL;
-
-                        if (_currentOffenseType != OffenseType.EVASION)
-                            _currentOffenseType = OffenseType.EVASION;
-                    }
-                }
-            };
-
-            _sturdyMachineControl.Deflection.Evasion.canceled += context =>
-            {
-                if (_isStanceActivated)
-                {
-                    if (_currentOffenseDirection != OffenseDirection.STANCE)
-                        _currentOffenseDirection = OffenseDirection.STANCE;
-                }
-            };
-        }
-
-        void OnEnable()
-        {
-            _sturdyMachineControl.Deflection.Enable();
-            _sturdyMachineControl.Sweep.Enable();
-            _sturdyMachineControl.Strikes.Enable();
-            _sturdyMachineControl.Heavy.Enable();
-            _sturdyMachineControl.DeathBlow.Enable();
-        }
-
-        void OnDisable()
-        {
-            _sturdyMachineControl.Deflection.Disable();
-            _sturdyMachineControl.Sweep.Disable();
-            _sturdyMachineControl.Strikes.Disable();
-            _sturdyMachineControl.Heavy.Disable();
-            _sturdyMachineControl.DeathBlow.Disable();
+            base.CustomLateUpdate(pOffenseDirection);
         }
 
         public override void OnCollisionEnter(Collision pCollision)
